alhamdulillah, successfully moved all of tagHolder to separate class file
